requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.2.8
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
command-not-found==0.3
ConfigArgParse==1.7
cryptography==3.4.8
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
dravid==0.7.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.95.1
fastjsonschema==2.19.1
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gevent==24.2.1
geventhttpclient==2.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib-metadata==4.6.4
inflate64==1.0.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
jiter==0.5.0
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
locust==2.25.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==8.10.0
msgpack==1.0.8
multidict==6.0.5
multivolumefile==0.2.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
notebook==7.2.0
notebook_shim==0.2.4
numpy==2.0.0
oauthlib==3.2.0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
Pigments==1.6
Pillow==9.5.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.45
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.21.1
pybcj==1.0.2
pycparser==2.22
pycryptodomex==3.20.0
pydantic==1.10.7
pydantic_core==2.20.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyppmd==1.1.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==5.4.1
pyzmq==25.1.2
pyzstd==0.16.0
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roundrobin==0.0.4
rpds-py==0.18.1
SecretStorage==3.3.1
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.26.1
systemd-python==234
terminado==0.18.1
texttable==1.7.0
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.22.0
wadllib==1.3.6
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.2
widgetsnbextension==4.0.11
xarray==2024.6.0
yarl==1.9.4
zipp==1.0.0
zope.event==5.0
zope.interface==6.3
